requirements.txt
numpy>=1.20.0
torch>=1.10.0
# For PyTorch Geometric, ensure your PyTorch version and CUDA setup (if any) are compatible.
# Refer to https://pytorch-geometric.readthedocs.io/en/latest/install/installation.html for specific versions.
# Example for PyTorch 2.3.0 and CUDA 12.1:
# torch_geometric==2.5.0
torch_geometric>=2.0.0
PyYAML>=6.0
tqdm>=4.62.0
# RDKit is often best installed via conda for full functionality.
# If installing via pip, ensure you have necessary system dependencies.
# pip install rdkit-pypi
rdkit-pypi>=2023.9.1
